Began Turn Order Images
Began working on the turn order image cycle. I have the start of a function to create the images and fill out the Canvas object which has a horizontal layout group. There's also a function that will set the first image to be the last and adjust the others accordingly by setting it as the last sibling.